// apps/frontend/src/types/BindingView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BindingTargetView } from "./BindingTargetView";

export type BindingView = { bindingId: string, skillIds: Array<string>, scope: string, projectRootId: string | null, agentIds: Array<string>, createdAt: string, targets: Array<BindingTargetView>, };